code/mqtt_client/requirements.txt
certifi==2021.5.30
idna==2.10
paho-mqtt==1.5.1
pytz==2021.1
requests==2.25.1
urllib3==1.26.5
